+++ app/models/cost_query/filter/author_id.rb
@@ -1,9 +1,9 @@
 class CostQuery::Filter::AuthorId < Report::Filter::Base
-  join_table Issue
-  applies_for :label_issue_attributes
+  join_table WorkPackage
+  applies_for :label_work_package_attributes
 
   def self.label
-    Issue.human_attribute_name(:author)
+    WorkPackage.human_attribute_name(:author)
   end
 
   def self.available_values(*)