+++ app/models/cost_query/group_by/author_id.rb
@@ -1,8 +1,8 @@
 class CostQuery::GroupBy::AuthorId < Report::GroupBy::Base
-  join_table Issue
-  applies_for :label_issue_attributes
+  join_table WorkPackage
+  applies_for :label_work_package_attributes
 
   def self.label
-    Issue.human_attribute_name(:author)
+    WorkPackage.human_attribute_name(:author)
   end
 end