+++ app/models/cost_query/group_by/user_id.rb
@@ -1,6 +1,6 @@
 class CostQuery::GroupBy::UserId < Report::GroupBy::Base
 
   def self.label
-    Issue.human_attribute_name(:user)
+    WorkPackage.human_attribute_name(:user)
   end
 end