+++ app/models/cost_query/group_by/work_package_id.rb
@@ -1,6 +1,6 @@
 class CostQuery::GroupBy::WorkPackageId < Report::GroupBy::Base
 
   def self.label
-    Issue.model_name.human
+    WorkPackage.model_name.human
   end
 end